+++ core/src/iAModality.h
@@ -1,7 +1,7 @@
 /*************************************  open_iA  ************************************ *
 * **********   A tool for visual analysis and processing of 3D CT images   ********** *
 * *********************************************************************************** *
-* Copyright (C) 2016-2019  C. Heinzl, M. Reiter, A. Reh, W. Li, M. Arikan, Ar. &  Al. *
+* Copyright (C) 2016-2020  C. Heinzl, M. Reiter, A. Reh, W. Li, M. Arikan, Ar. &  Al. *
 *                          Amirkhanov, J. Weissenböck, B. Fröhler, M. Schiwarth       *
 * *********************************************************************************** *
 * This program is free software: you can redistribute it and/or modify it under the   *
@@ -61,9 +61,9 @@ public:
 	//! return the channel in the specified file that the data in this class comes from (don't confuse with channelID!)
 	int channel() const;
 	//! return the number of components in this modality
-	int componentCount() const;
+	size_t componentCount() const;
 	//! return a specific component of this modality
-	vtkSmartPointer<vtkImageData> component(int idx) const;
+	vtkSmartPointer<vtkImageData> component(size_t idx) const;
 	//! get the name of the transfer function file
 	QString transferFileName() const;
 	//! set name of the modality
